requirement.txt
flask
python-decouple
pip install psycopg2